chore: move screenshot to its own place, add chrono for timestamp files, latest shaders
