+++ Cargo.toml
@@ -17,6 +17,7 @@ bevy = { version= "0.11.*", features = [
 ] }
 
 bevy_panorbit_camera = "0.8.*"
+chrono = "0.4.31"
 
 [build-dependencies]
 glob = "0.3.1"
