+++ assets/shaders/myshader.wgsl
@@ -3,37 +3,19 @@
 
 const TAU:f32 =  6.28318530718;
 
-fn plot(st: vec2f, pct: f32) -> f32 {
-    let l = pct - 0.02;
-    let r = pct + 0.02;
-
-    return smoothstep(l, pct, st.y) - smoothstep(pct, r, st.y);
-}
-
 @fragment
 fn fragment(in: MeshVertexOutput) -> @location(0) vec4<f32> {
+    // let uv = in.uv;
     var uv = (in.uv * 2.0) - 1.0;
-    // var uv = in.uv;
     var col = vec3f(0.);
-    uv = fract(uv) * 1.0;
 
-    let toCenter = vec2(0.25) - uv;
-    let angle = atan2(toCenter.y, toCenter.x);
-    let radius = length(toCenter) * 2.0;
+    let distance_to_center = vec2(0.5) - uv;
+    let angle = atan2(distance_to_center.y, distance_to_center.x);
+    let radius = length(distance_to_center) * 2.0;
 
     col = hsv_to_srgb(vec3f((angle / TAU) + globals.time / 3.0, radius, 1.0));
 
-    let circ = circle(uv, 0.6);
-    col *= circ;
-
-    let pct = distance(uv, vec2f(0.5));
-
     return vec4f(col, 1.0);
-}
-
-fn circle(st: vec2f, rad: f32) -> f32 {
-    let dist = st - vec2f(0.5);
-    return 1.0 - smoothstep(rad - (rad * 0.01), rad + (rad * 0.01), dot(dist, dist) * 4.0);
 }
 
 // From the bevy source code
