+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod screenshot;
 pub mod shader_utils;
 pub mod ui;
 pub mod utils;
